feat: CardType 기능 추가
